Move telemetry utils into its dedicated folder in ./utils
