+++ docs/edit/runbook.md
@@ -20,8 +20,8 @@ The impact is that these **configs are not visible** in Metabase, REDAPL, or any
 2. Add config normalization rules [here](https://github.com/DataDog/dd-go/tree/prod/trace/apps/tracer-telemetry-intake/telemetry-payload/static/) following the existing pattern
    1. This can be merged with any review from [@apm-sdk](https://github.com/orgs/DataDog/teams/apm-sdk)
    2. Bonus Points: Run the auto-formatter [_format.py](https://github.com/DataDog/dd-go/blob/prod/trace/apps/tracer-telemetry-intake/telemetry-payload/static/_format.py) from the `dd-go` root via `python ./trace/apps/tracer-telemetry-intake/telemetry-payload/static/_format.py`
-3. After merging, update system-tests by running [update.sh](/tests/telemetry_intake/update.sh)
-   1. This can be run from the root by running `./tests/telemetry_intake/update.sh`
+3. After merging, update system-tests by running [update.sh](/utils/telemetry/intake/update.sh)
+   1. This can be run from the root by running `./utils/telemetry/intake/update.sh`
    2. This can be merged with any review from [@apm-ecosystems](https://github.com/orgs/DataDog/teams/apm-ecosystems)
 4. You're all set - your tests should pass 🏁
 
@@ -34,9 +34,9 @@ tests.test_telemetry.test_config_telemetry_completeness
 
 This asserts that config telemetry is handled properly by telemetry intake
 
-Some files are manually copied from dd-go from/to the following paths using tests/telemetry_intake/update.sh
+Some files are manually copied from dd-go from/to the following paths using utils/telemetry/intake/update.sh
 from: https://github.com/DataDog/dd-go/blob/prod/trace/apps/tracer-telemetry-intake/telemetry-payload/static/
-to: tests/telemetry_intake/static
+to: utils/telemetry/intake/static
 
 If this test fails, it means that a telemetry key was found in config telemetry that does not
 exist in any of the files listed above in dd-go
@@ -51,4 +51,4 @@ To fix this, you must update dd-go to either
 Once dd-go is updated, you can copy over the files to this repo and merge them in as part of your changes
 This can be done by running the following from the src root
 
-Usage: ./tests/telemetry_intake/update.sh
+Usage: ./utils/telemetry/intake/update.sh
