+++ utils/telemetry/intake/static/config_aggregation_list.json
