+++ utils/telemetry/intake/static/config_norm_rules.json
