+++ utils/telemetry/intake/static/dotnet_config_rules.json
