+++ utils/telemetry/intake/static/go_config_rules.json
