+++ utils/telemetry/intake/static/jvm_config_rules.json
