+++ utils/telemetry/intake/static/nodejs_config_rules.json
