+++ utils/telemetry/intake/static/php_config_rules.json
