+++ utils/telemetry/intake/static/python_config_rules.json
