+++ utils/telemetry/intake/static/ruby_config_rules.json
